Refactor tweetbot.py to handle all configurations in a single command
This commit refactors the tweetbot.py file to handle all configurations in a single command called "all_in_one". Previously, there were separate commands for single configuration and multiple configurations. The "all_in_one" command allows users to send a list of configurations and send them with an ad. The handle_message function has been modified to accommodate this change, with the addition of a new parameter "divide" to control whether the configurations should be divided or not. This change improves the user experience by simplifying the configuration process.